Merge SiteWhere 3.0 beta code to master.

+++ service-event-sources/src/main/java/com/sitewhere/sources/decoder/protobuf/ProtobufDeviceEventEncoder.java
@@ -28,8 +28,6 @@ import com.sitewhere.spi.device.event.request.IDeviceRegistrationRequest;
 /**
  * Implementation of {@link IDeviceEventEncoder} that encodes device events into
  * binary using the SiteWhere Google Protocol Buffers format.
- * 
- * @author Derek
  */
 public class ProtobufDeviceEventEncoder implements IDeviceEventEncoder<byte[]> {
 
@@ -74,10 +72,11 @@ public class ProtobufDeviceEventEncoder implements IDeviceEventEncoder<byte[]> {
 	    if (measurements.getMetadata() != null) {
 		payloadBuilder.putAllMetadata(measurements.getMetadata());
 	    }
-	    
+
 	    payloadBuilder.setMeasurementName(GOptionalString.newBuilder().setValue(measurements.getName()));
-	    payloadBuilder.setMeasurementValue(GOptionalDouble.newBuilder().setValue(measurements.getValue()));
-	    
+	    payloadBuilder
+		    .setMeasurementValue(GOptionalDouble.newBuilder().setValue(measurements.getValue().doubleValue()));
+
 	    // Write to byte-stream
 	    ByteArrayOutputStream out = new ByteArrayOutputStream();
 	    headerBuilder.build().writeDelimitedTo(out);
@@ -141,9 +140,9 @@ public class ProtobufDeviceEventEncoder implements IDeviceEventEncoder<byte[]> {
 	    // Payload
 	    DeviceEvent.DeviceLocation.Builder payloadBuilder = DeviceEvent.DeviceLocation.newBuilder();
 	    payloadBuilder.setEventDate(GOptionalFixed64.newBuilder().setValue(location.getEventDate().getTime()));
-	    payloadBuilder.setLatitude(GOptionalDouble.newBuilder().setValue(location.getLatitude()));
-	    payloadBuilder.setLongitude(GOptionalDouble.newBuilder().setValue(location.getLongitude()));
-	    payloadBuilder.setElevation(GOptionalDouble.newBuilder().setValue(location.getElevation()));
+	    payloadBuilder.setLatitude(GOptionalDouble.newBuilder().setValue(location.getLatitude().doubleValue()));
+	    payloadBuilder.setLongitude(GOptionalDouble.newBuilder().setValue(location.getLongitude().doubleValue()));
+	    payloadBuilder.setElevation(GOptionalDouble.newBuilder().setValue(location.getElevation().doubleValue()));
 	    if (location.getMetadata() != null) {
 		payloadBuilder.putAllMetadata(location.getMetadata());
 	    }
